alteracao de configuracoes de campos do banco

+++ database/migrations/2024_04_09_213009_create_transacoes_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('transacoes', function (Blueprint $table) {
             $table->id();
             $table->integer('conta_id')->nullable(false);
-            $table->integer('valor')->nullable(false);
+            $table->decimal('valor',12,2)->nullable(false);
             $table->timestamps();
         });
     }
